requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosmtplib==2.0.2
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
alembic==1.12.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.16
click==8.1.7
cloudinary==1.36.0
coloredlogs==15.0.1
dataclasses-json==0.6.1
Deprecated==1.2.14
distro==1.8.0
dnspython==2.4.2
document==1.0
ecdsa==0.18.0
email-validator==2.1.0.post1
environ==1.0
faiss-cpu==1.7.4
fastapi==0.104.1
fastapi-limiter==0.1.5
fastapi-mail==1.4.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kubernetes==28.1.0
langchain==0.0.332
langsmith==0.0.62
lark==1.1.8
libgravatar==1.0.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.16.1
openai==0.28.1
overrides==7.4.0
packaging==23.2
pandas==2.1.2
passlib==1.7.4
Pillow==10.1.0
pluggy==1.3.0
posthog==3.0.2
protobuf==4.25.0
psycopg2-binary==2.9.9
pulsar-client==3.3.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0

pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pypdf==3.17.0
PyPDF2==3.0.1
PyPika==0.48.9
pytest==7.4.3
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.28.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
wrapt==1.16.0
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.17.0
